requirements.txt
PyQt5~=5.15.10
python-can~=4.3.1
pymongo~=4.6.1
shortuuid~=1.0.11